Add images folder + edit both HTML files
an image folder was created to store the images used in the recipes.
The title of the recipe was changed, both on the index and inside
the recipe file itself.
A "go back" link to the homepage was added to the recipe file.
The link in the index file no longer forces a new tab to be opened.