ADD CSS + Remove Go Back Button
A stylesheet was added in the appropriate folder and linked to
all three webpages. In order to properly select different elements,
<div> containers were used along with class and id attributes: the
ingredients and steps section were transformed into sibling elements
and then floated next to each other for a cleaner look.

Font families used were Cambria and Helvetica, using system-ui as
fallback. Images were cropped into circular shapes.

The Go Back button was removed because the appearance could not be
standardised along all three webpages. It was originally changed to
appear at the bottom of the page, after all other sections, but it on
two pages, it was showing up right below the ingredients section and
to the left of the steps section, in spite of it being outside the div
containers.