Update CSS + Reintroduce Home Button
The styling of the elements was modified: both the image and description
sections were brought into a div container and reorganised so that
the image would show on the top left of all webpages, below the
heading, whilst the description directly to its right. The ingredients
and steps sections show right below them.

A Home button was successfully implemented at the bottom of each recipe
page.